feat(frontend): add vue router navigation
