+++ frontend/src/main.js
@@ -1,6 +1,11 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+import router from './router'
 
 import './assets/base.css'
 
-createApp(App).mount('#app')
+const app = createApp(App)
+
+app.use(router)
+
+app.mount('#app')
